requirements.txt
uvicorn[standard]==0.15.0
gunicorn==20.1.0
fastapi==0.68.1
pydantic>=1.8.0,<2.0.0
fastapi-socketio==0.0.9